+++ wp-content/plugins/optinmonster/OMAPI/WooCommerce.php
@@ -73,6 +73,10 @@ class OMAPI_WooCommerce {
 
 		// Add custom OptinMonster note.
 		add_action( 'admin_init', array( $this, 'maybe_store_note' ) );
+
+		// Revenue attribution support.
+		add_action( 'woocommerce_thankyou', array( $this, 'maybe_store_revenue_attribution' ) );
+		add_action( 'woocommerce_order_status_changed', array( $this, 'maybe_store_revenue_attribution_on_order_status_change' ), 10, 3 );
 	}
 
 	/**
@@ -613,5 +617,80 @@ class OMAPI_WooCommerce {
 		);
 
 		$note->save();
+	}
+
+	/**
+	 * Maybe stores revenue attribution data when a purchase is successful.
+	 *
+	 * @since 2.6.13
+	 *
+	 * @param int  $order_id The WooCommerce order ID.
+	 * @param bool $force    Flag to force storing the revenue attribution data.
+	 *
+	 * @return void
+	 */
+	public function maybe_store_revenue_attribution( $order_id = 0, $force = false ) {
+		// If we have already stored revenue attribution data before, return early.
+		$stored = get_post_meta( $order_id, '_om_revenue_attribution_complete', true );
+		if ( $stored ) {
+			return;
+		}
+
+		// Grab the order. If we can't, return early.
+		$order = wc_get_order( $order_id );
+		if ( ! $order ) {
+			return;
+		}
+
+		// Grab some necessary data to send.
+		$data_on_order = get_post_meta( $order_id, '_om_revenue_attribution_data', true );
+		$data          = wp_parse_args(
+			array(
+				'transaction_id' => absint( $order_id ),
+				'value'          => esc_html( $order->get_total() ),
+			),
+			! empty( $data_on_order ) ? $data_on_order : $this->base->revenue->get_revenue_data()
+		);
+
+		// If the order is not complete, return early.
+		// This will happen for payments where further
+		// work is required (such as checks, etc.). In those
+		// instances, we need to store the data to be processed
+		// at a later time.
+		if ( ! $order->has_status( 'completed' ) && ! $force ) {
+			update_post_meta( $order_id, '_om_revenue_attribution_data', $data );
+			return;
+		}
+
+		// Attempt to make the revenue attribution request.
+		// It checks to determine if campaigns are set, etc.
+		$ret = $this->base->revenue->store( $data );
+		if ( ! $ret || is_wp_error( $ret ) ) {
+			return;
+		}
+
+		// Update the payment meta for storing revenue attribution data.
+		update_post_meta( $order_id, '_om_revenue_attribution_complete', time() );
+	}
+
+	/**
+	 * Maybe stores revenue attribution data when a purchase is successful.
+	 *
+	 * @since 2.6.13
+	 *
+	 * @param int    $order_id   The WooCommerce order ID.
+	 * @param string $old_status The old order status.
+	 * @param string $new_status The new order status.
+	 *
+	 * @return void
+	 */
+	public function maybe_store_revenue_attribution_on_order_status_change( $order_id, $old_status, $new_status ) {
+		// If we don't have the proper new status, return early.
+		if ( 'completed' !== $new_status ) {
+			return;
+		}
+
+		// Maybe store the revenue attribution data.
+		return $this->maybe_store_revenue_attribution( $order_id, true );
 	}
 }