+++ wp-content/plugins/optinmonster/optin-monster-wp-api.php
@@ -5,7 +5,7 @@
  * Description: OptinMonster is the best WordPress popup builder plugin that helps you grow your email newsletter list and sales with email popups, exit intent popups, floating bars and more!
  * Author:      OptinMonster Popup Builder Team
  * Author URI:  https://optinmonster.com
- * Version:     2.6.12
+ * Version:     2.7.0
  * Text Domain: optin-monster-api
  * Domain Path: languages
  *
@@ -13,7 +13,7 @@
  * WC tested up to:      6.2.0
  * Requires at least:    4.7.0
  * Requires PHP:         5.3
- * Tested up to:         5.9.2
+ * Tested up to:         5.9.3
  *
  * OptinMonster is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -66,7 +66,7 @@ class OMAPI {
 	 *
 	 * @var string
 	 */
-	public $version = '2.6.12';
+	public $version = '2.7.0';
 
 	/**
 	 * The name of the plugin.
@@ -416,9 +416,11 @@ class OMAPI {
 		$this->type        = new OMAPI_Type();
 		$this->output      = new OMAPI_Output();
 		$this->shortcode   = new OMAPI_Shortcode();
+		$this->revenue     = new OMAPI_RevenueAttribution();
 		$this->woocommerce = new OMAPI_WooCommerce();
 		$this->elementor   = new OMAPI_Elementor();
 		$this->mailpoet    = new OMAPI_MailPoet();
+		$this->edd         = new OMAPI_EasyDigitalDownloads();
 
 		if ( defined( 'DOING_CRON' ) && DOING_CRON && ! $this->actions ) {
 			$this->save    = new OMAPI_Save();
@@ -971,6 +973,19 @@ class OMAPI {
 	public function get_site_id() {
 		$option = $this->get_option();
 		return ! empty( $option['siteId'] ) ? (string) $option['siteId'] : '';
+	}
+
+	/**
+	 * Gets the revenue attribution settings associated with the
+	 * account that has registered this site.
+	 *
+	 * @since 2.6.13
+	 *
+	 * @return array
+	 */
+	public function get_revenue_attribution() {
+		$option = $this->get_option();
+		return ! empty( $option['revenueAttribution'] ) ? (array) $option['revenueAttribution'] : array();
 	}
 
 	/**